Maya/229189 sample table pathogen config (#1597)
* move sample table column defs into their own files

* configure table columns by pathogen

* types

* different def for publicid col sc2 vs mpox

* types
