+++ src/frontend/src/views/Data/components/SamplesView/components/SamplesTable/columnDefinitions/style.ts
@@ -1,13 +1,15 @@
 import styled from "@emotion/styled";
 import { CellBasic, getFontWeights } from "czifui";
 
-export const StyledCellBasic = styled(CellBasic)`
+export const StyledPrivateId = styled(CellBasic)`
+  padding-left: 0;
+
   ${(props) => {
     const fontWeights = getFontWeights(props);
     return `
-      span {
+      span:first-of-type {
         font-weight: ${fontWeights?.semibold};
-        word-break: break-word;
+        word-break: break-all;
       }
     `;
   }}
